[fix] attributes fixed

+++ hackathon/src/main/java/com/hacker/hackathon/model/UserTodoQuiz.java
@@ -18,7 +18,7 @@ public class UserTodoQuiz {
     private Long userTodoQuizId;
 
     @Column
-    private int stage;
+    private Long stage;
 
     @Column
     private String previousAnswer;
